experiments/chatBOT/requirements.txt
flask
requests
beautifulsoup4
sentence-transformers
chromadb
google-generativeai
langchain-text-splitters
python-dotenv
groq